gcc signed/unsigned comparison fix
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@42792 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/msw/printdlg.cpp
@@ -200,7 +200,7 @@ bool wxWindowsPrintNativeData::TransferTo( wxPrintData &data )
         }
         if (devMode->dmFields & DM_MEDIATYPE)
         {
-            wxASSERT(devMode->dmMediaType != wxPRINTMEDIA_DEFAULT);
+            wxASSERT( (int)devMode->dmMediaType != wxPRINTMEDIA_DEFAULT );
             data.SetMedia(devMode->dmMediaType);
         }
         //// Printer name
